+++ src/core/utils/math.js
@@ -57,38 +57,6 @@ export function isCircularPath(path) {
 	return path.type === 'circle' || path.type === 'oval';
 }
 
-export function decodePolyline(encoded) {
-	let len = encoded.length;
-	let index = 0;
-	let lat = 0;
-	let lng = 0;
-	let array = [];
-
-	while (index < len) {
-		let b;
-		let shift = 0;
-		let result = 0;
-		do {
-			b = encoded.charCodeAt(index++) - 63;
-			result |= (b & 0x1f) << shift;
-			shift += 5;
-		} while (b >= 0x20);
-		let dlat = ((result & 1) ? ~(result >> 1) : (result >> 1));
-		lat += dlat;
-		shift = 0;
-		result = 0;
-		do {
-			b = encoded.charCodeAt(index++) - 63;
-			result |= (b & 0x1f) << shift;
-			shift += 5;
-		} while (b >= 0x20);
-		let dlng = ((result & 1) ? ~(result >> 1) : (result >> 1));
-		lng += dlng;
-		array.push([lat * 1e-5, lng * 1e-5]);
-	}
-	return array;
-}
-
 export function mapValue(value, inMin, inMax, outMin, outMax) {
 	const dividend = outMax - outMin;
 	const divisor = inMax - inMin;